# docs/schema/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "Animateur" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "prenom" : {
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "dateNaissance" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "manager" : {
            "type" : "boolean"
          },
          "email" : {
            "type" : "string"
          },
          "accessToken" : {
            "type" : "string"
          },
          "competences" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/NiveauCompetence"
            }
          },
          "souhaits" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          },
          "joursIndisponibles" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string",
              "format" : "date",
              "examples" : [ "2022-03-10" ]
            }
          },
          "ninja" : {
            "type" : "boolean"
          },
          "modifieLe" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "ConstraintToggle" : {
        "type" : "object",
        "properties" : {
          "nom" : {
            "type" : "string"
          }
        }
      },
...
      "ContrainteAdHoc" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/TypeContrainteAdHoc"
          },
          "animateursConcernes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Animateur"
            }
          },
          "creneau" : {
            "$ref" : "#/components/schemas/Creneau"
          },
          "stand" : {
            "$ref" : "#/components/schemas/Stand"
          },
          "raison" : {
            "type" : "string"
          },
          "creeParUtilisateurId" : {
            "type" : "string"
          },
          "creeLe" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "modifieLe" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "Creneau" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "jour" : {
            "type" : "integer",
            "format" : "int32"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "famille" : {
            "type" : "integer",
            "format" : "int32"
          },
          "couverturePause" : {
            "type" : "boolean"
          },
          "modifieLe" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "dureeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "DayOfWeek" : {
        "type" : "string",
        "enum" : [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
      },
...
      "Emplacement" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "latitude" : {
            "type" : "number",
            "format" : "double"
          },
          "longitude" : {
            "type" : "number",
            "format" : "double"
          },
          "modifieLe" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "FenetreHoraire" : {
        "type" : "object",
        "properties" : {
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "effectif" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "HardMediumSoftScore" : {
        "type" : "object",
        "properties" : {
          "zero" : {
            "type" : "boolean"
          },
          "hardScore" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mediumScore" : {
            "type" : "integer",
            "format" : "int64"
          },
          "softScore" : {
            "type" : "integer",
            "format" : "int64"
          },
          "feasible" : {
            "type" : "boolean"
          }
        }
      },
...
      "HoraireStand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mode" : {
            "$ref" : "#/components/schemas/ModeHoraire"
          },
          "jours" : {
            "$ref" : "#/components/schemas/TypeJoursHoraire"
          },
          "joursSemaine" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "$ref" : "#/components/schemas/DayOfWeek"
            }
          },
          "dateDebut" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "dateFin" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "dates" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string",
              "format" : "date",
              "examples" : [ "2022-03-10" ]
            }
          },
          "fenetres" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FenetreHoraire"
            }
          },
          "motif" : {
            "type" : "string"
          }
        }
      },
...
      "ImportTargetView" : {
        "required" : [ "existe" ],
        "type" : "object",
        "properties" : {
          "editionId" : {
            "type" : "string"
          },
          "editionNomFichier" : {
            "type" : "string"
          },
          "existe" : {
            "type" : "boolean"
          },
          "editionNomExistant" : {
            "type" : "string"
          }
        }
      },
...
      "IndisponibiliteStand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "motif" : {
            "type" : "string"
          }
        }
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
...
      "LocalTime" : {
        "type" : "string",
        "format" : "local-time",
        "examples" : [ "13:45:30.123456789" ],
        "externalDocs" : {
          "description" : "As defined by 'partial-time' in RFC3339",
          "url" : "https://www.rfc-editor.org/rfc/rfc3339.html#section-5.6"
        }
      },
...
      "ModeHoraire" : {
        "type" : "string",
        "enum" : [ "OUVERTURE", "FERMETURE" ]
      },
...
      "NiveauCompetence" : {
        "type" : "string",
        "enum" : [ "DEBUTANT", "AUTONOME", "REFERENT" ]
      },
      "NiveauEffort" : {
        "type" : "string",
        "enum" : [ "NORMAL", "EPUISANT" ]
      },
...
      "OuvertureStand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "motif" : {
            "type" : "string"
          },
          "effectif" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ParametresLegaux" : {
        "type" : "object",
        "properties" : {
          "dureeHebdomadaireMaxMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dureeHebdomadaireMaxMineurMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pauseMinimaleEntreVacationsMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reposQuotidienMinimalMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pauseSurPoste" : {
            "type" : "boolean"
          }
        }
      },
...
      "ParametresQualite" : {
        "required" : [ "maxEmplacementsDistinctsParJour" ],
        "type" : "object",
        "properties" : {
          "maxEmplacementsDistinctsParJour" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PlanningEvenement" : {
        "type" : "object",
        "properties" : {
          "dateDebutFestival" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "animateurs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Animateur"
            }
          },
          "postes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PosteAffectation"
            }
          },
          "contraintesAdHoc" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ContrainteAdHoc"
            }
          },
          "parametresLegaux" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ParametresLegaux"
            }
          },
          "parametresQualite" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ParametresQualite"
            }
          },
          "constraintsDesactivees" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConstraintToggle"
            }
          },
          "verrouillages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VerrouillagePlanning"
            }
          },
          "score" : {
            "$ref" : "#/components/schemas/HardMediumSoftScore"
          }
        }
      },
...
      "PosteAffectation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "stand" : {
            "$ref" : "#/components/schemas/Stand"
          },
          "creneau" : {
            "$ref" : "#/components/schemas/Creneau"
          },
          "heureDebutEffective" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFinEffective" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "animateur" : {
            "$ref" : "#/components/schemas/Animateur"
          },
          "verrouille" : {
            "type" : "boolean"
          },
          "dureeEffectiveMinutes" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ScenarioValidationResult" : {
        "required" : [ "valide" ],
        "type" : "object",
        "properties" : {
          "valide" : {
            "type" : "boolean"
          },
          "erreurs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "Stand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "typologiesProposees" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          },
          "effectifMin" : {
            "type" : "integer",
            "format" : "int32"
          },
          "effectifMax" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reserveMajeurs" : {
            "type" : "boolean"
          },
          "premium" : {
            "type" : "boolean"
          },
          "famille" : {
            "type" : "integer",
            "format" : "int32"
          },
          "niveauEffort" : {
            "$ref" : "#/components/schemas/NiveauEffort"
          },
          "emplacement" : {
            "$ref" : "#/components/schemas/Emplacement"
          },
          "indisponibilites" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IndisponibiliteStand"
            }
          },
          "ouvertures" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OuvertureStand"
            }
          },
          "horaires" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/HoraireStand"
            }
          },
          "modifieLe" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "TypeContrainteAdHoc" : {
        "type" : "string",
        "enum" : [ "INDISPONIBILITE_FORCEE", "INCOMPATIBILITE", "AFFECTATION_FORCEE", "AFFINITE" ]
      },
      "TypeJoursHoraire" : {
        "type" : "string",
        "enum" : [ "TOUS", "JOURS_SEMAINE", "PLAGE", "DATES" ]
      },
      "TypeVerrouillage" : {
        "type" : "string",
        "enum" : [ "ANIMATEUR", "STAND", "JOUR", "CRENEAU", "ANIMATEUR_CRENEAU" ]
      },
...
      "VerrouillagePlanning" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/TypeVerrouillage"
          },
          "animateurId" : {
            "type" : "string"
          },
          "standId" : {
            "type" : "string"
          },
          "creneauId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "jour" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "raison" : {
            "type" : "string"
          },
          "creeLe" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
    }
  },
  "paths" : {
...
    "/api/reference-data/cible-scenario" : {
      "get" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportTargetView"
                }
              }
            }
          }
        },
        "summary" : "Scenario Target",
        "tags" : [ "Reference Data Resource" ]
      }
    },
    "/api/reference-data/cible-scenario-fichier" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "File Scenario Target",
        "tags" : [ "Reference Data Resource" ]
      }
    },
...
    "/api/reference-data/import" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlanningEvenement"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Import Reference Data",
        "tags" : [ "Reference Data Resource" ]
      }
    },
    "/api/reference-data/import-scenario" : {
      "post" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Import Scenario",
        "tags" : [ "Reference Data Resource" ]
      }
    },
    "/api/reference-data/import-scenario-fichier" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Import Scenario File",
        "tags" : [ "Reference Data Resource" ]
      }
    },
    "/api/reference-data/valider-scenario-fichier" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScenarioValidationResult"
                }
              }
            }
          }
        },
        "summary" : "Validate Scenario File",
        "tags" : [ "Reference Data Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "planning-equipes API",
    "version" : "999-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}